# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Smart Access Monitor API",
    "version": "1.0.0",
    "description": "JWT login for dashboard reads; per-device API key for Pi writes. All timestamps are server-generated. Import this file into Swagger Editor or Postman. Cookies are also accepted for dashboard reads."
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "deviceKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "smart_access_token"
      }
    }
  },
  "paths": {
...
    "/api/devices": {
      "get": {
        "summary": "List registered devices without credential hashes",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "example": [
                  {
                    "id": 1,
                    "device_uid": "main-entrance-pi",
                    "name": "Main entrance Raspberry Pi",
                    "location": null,
                    "last_seen": "2026-10-03T06:00:00Z",
                    "device_online": true
                  }
                ]
              }
            }
          },
          "400": {
            "description": "Bad input",
            "content": {
              "application/json": {
                "example": {
                  "success": false,
                  "message": "Invalid request"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid credentials",
            "content": {
              "application/json": {
                "example": {
                  "success": false,
                  "message": "Login required"
                }
              }
            }
          },
          "500": {
            "description": "Server/database failure; retry later",
            "content": {
              "application/json": {
                "example": {
                  "success": false,
                  "message": "Internal server error"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ],
        "parameters": []
      }
    },
...
    "/api/device/verify-access": {
      "post": {
        "summary": "Verify the enrolled RFID card and PIN together; either may be entered first on the Pi",
        "security": [
          {
            "deviceKey": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["device_uid", "card_uid", "pin"],
                "properties": {
                  "device_uid": {"type": "string", "maxLength": 100},
                  "card_uid": {"type": "string", "maxLength": 64},
                  "pin": {"type": "string", "pattern": "^[0-9]{4,8}$"}
                }
              },
              "example": {
                "device_uid": "main-entrance-pi",
                "card_uid": "<card UID>",
                "pin": "1234"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Pair checked; invalid factors receive a generic denial",
            "content": {
              "application/json": {
                "examples": {
                  "granted": {"value": {"granted": true, "reason": "authorized", "user_id": 1, "card_id": 1, "name": "Example User"}},
                  "denied": {"value": {"granted": false, "reason": "invalid_credentials"}}
                }
              }
            }
          },
          "400": {"description": "Invalid request body"},
          "401": {"description": "Missing or invalid device API key"},
          "429": {"description": "PIN attempts rate-limited"},
          "500": {"description": "Server/database failure; access must remain denied"}
        },
        "description": "The raw PIN is checked against a bcrypt hash and the card UID against its SHA-256 hash. The Pi must not grant access unless this endpoint returns granted=true."
      }
    },
    "/api/access": {
      "post": {
        "summary": "Record a Pi-checked RFID/PIN attempt",
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "example": {
                  "success": true,
                  "message": "Access attempt stored"
                }
              }
            }
          },
          "400": {
            "description": "Bad input",
            "content": {
              "application/json": {
                "example": {
                  "success": false,
                  "message": "Invalid request"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid credentials",
            "content": {
              "application/json": {
                "example": {
                  "success": false,
                  "message": "Login required"
                }
              }
            }
          },
          "500": {
            "description": "Server/database failure; retry later",
            "content": {
              "application/json": {
                "example": {
                  "success": false,
                  "message": "Internal server error"
                }
              }
            }
          }
        },
        "security": [
          {
            "deviceKey": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              },
              "example": {
                "device_uid": "main-entrance-pi",
                "method": "pin",
                "result": "denied",
                "reason": "wrong_pin"
              }
            }
          }
        },
        "description": "Required method: rfid or pin; result: granted or denied. Optional reason (max 50 chars), user_id, card_id, snapshot_id (existing positive integer IDs or null). Pi performs access checking. No PINs or raw card UIDs are sent here."
      },
      "get": {
        "summary": "Recent access attempts, newest first (aliases)",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "example": [
                  {
                    "id": "1",
                    "device_uid": "main-entrance-pi",
                    "created_at": "2026-10-03T06:00:00Z",
                    "method": "pin",
                    "result": "denied",
                    "reason": "wrong_pin",
                    "user_id": null,
                    "user_name": null,
                    "card_id": null,
                    "snapshot_id": null
                  }
                ]
              }
            }
          },
          "400": {
            "description": "Bad input",
            "content": {
              "application/json": {
                "example": {
                  "success": false,
                  "message": "Invalid request"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid credentials",
            "content": {
              "application/json": {
                "example": {
                  "success": false,
                  "message": "Login required"
                }
              }
            }
          },
          "500": {
            "description": "Server/database failure; retry later",
            "content": {
              "application/json": {
                "example": {
                  "success": false,
                  "message": "Internal server error"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "device_uid",
            "in": "query",
            "description": "Filter by registered device UID. Omit to query all devices (latest device for current KPI).",
            "schema": {
              "type": "string"
            },
            "example": "main-entrance-pi"
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 1000,
              "default": 50
            },
            "example": 50
          }
        ]
      }
    },
    "/api/access-logs": {
      "get": {
        "summary": "Recent access attempts, newest first (aliases)",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "example": [
                  {
                    "id": "1",
                    "device_uid": "main-entrance-pi",
                    "created_at": "2026-10-03T06:00:00Z",
                    "method": "pin",
                    "result": "denied",
                    "reason": "wrong_pin",
                    "user_id": null,
                    "user_name": null,
                    "card_id": null,
                    "snapshot_id": null
                  }
                ]
              }
            }
          },
          "400": {
            "description": "Bad input",
            "content": {
              "application/json": {
                "example": {
                  "success": false,
                  "message": "Invalid request"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid credentials",
            "content": {
              "application/json": {
                "example": {
                  "success": false,
                  "message": "Login required"
                }
              }
            }
          },
          "500": {
            "description": "Server/database failure; retry later",
            "content": {
              "application/json": {
                "example": {
                  "success": false,
                  "message": "Internal server error"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "device_uid",
            "in": "query",
            "description": "Filter by registered device UID. Omit to query all devices (latest device for current KPI).",
            "schema": {
              "type": "string"
            },
            "example": "main-entrance-pi"
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 1000,
              "default": 50
            },
            "example": 50
          }
        ]
      }
    },
...
  }
}